Fixing merging problems + goimports
